Resolve exo player issue and working file with local file.

+++ android/src/main/java/com/reactnativeaudiowaveform/visualizer/Utils.kt
@@ -6,7 +6,7 @@ import android.util.TypedValue
 object Utils {
 
     @JvmStatic
-    fun dp(context: Context?, dp: Int): Float {
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp.toFloat(), context!!.resources.displayMetrics)
+    fun dp(context: Context?, dp: Float): Float {
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, context!!.resources.displayMetrics)
     }
 }
